+++ openeo_driver/views.py
@@ -42,7 +42,6 @@ from openeo_driver.backend import (
     OpenEoBackendImplementation,
     ServiceMetadata,
     UserDefinedProcessMetadata,
-    function_has_argument,
     is_not_implemented,
     QueryablesListing,
 )
@@ -75,6 +74,7 @@ from openeo_driver.processgraph import ProcessGraphFlatDict, extract_default_job
 from openeo_driver.save_result import SaveResult, to_save_result
 from openeo_driver.users import User, user_id_b64_decode, user_id_b64_encode
 from openeo_driver.users.auth import HttpAuthHandler
+from openeo_driver.util.compat import function_has_argument
 from openeo_driver.util.geometry import BoundingBox, reproject_geometry
 from openeo_driver.util.logging import ExtraLoggingFilter, FlaskRequestCorrelationIdLogging
 from openeo_driver.util.stac import sniff_stac_extension_prefix